Unify branches and resolve conflicts
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/danvoulez/minimay?shareId=XXXX-XXXX-XXXX-XXXX).
