+++ Makefile
@@ -2,7 +2,7 @@ install:
 	bash install.sh
 
 run:
-	flask --app backend.app run
+	flask --app backend.app run && cd frontend && npm run build
 
 sync:
 	python backend/sync_fallback.py
