--- frontend/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "flipapp",
-  "version": "0.0.1",
-  "scripts": {
-    "dev": "svelte-kit dev",
-    "build": "svelte-kit build",
-    "preview": "svelte-kit preview"
-  },
-  "devDependencies": {
-    "@sveltejs/kit": "next",
-    "svelte": "^3.0.0",
-    "vite": "^4.0.0",
-    "tailwindcss": "^3.3.0",
-    "postcss": "^8.4.0",
-    "autoprefixer": "^10.4.0",
-    "@motionone/svelte": "^10.14.0",
-    "@supabase/supabase-js": "^2.0.0"
-  }
-}
