--- frontend/static/prompts.extended 2.json
@@ -1,5 +1,0 @@
-[
-  "Evento iniciado",
-  "Erro detectado",
-  "Processo concluido"
-]
