--- frontend/svelte.config.js
@@ -1,7 +1,0 @@
-import adapter from '@sveltejs/adapter-auto';
-
-export default {
-  kit: {
-    adapter: adapter()
-  }
-};
