--- frontend/vite.config.js
@@ -1,5 +1,0 @@
-import { sveltekit } from '@sveltejs/kit/vite';
-
-export default {
-  plugins: [sveltekit()]
-};
